requirements.txt
PySide2~=5.15.2
requests~=2.26.0
PyGithub~=1.55
ruamel.yaml